pyproject.toml
[build-system]
requires = ["setuptools>=75.3.0"]
build-backend = "setuptools.build_meta"

[project]
name = "forbidden"
version = "13.0"
authors = [{ name = "Ivan Sincek" }]
description = "Bypass 4xx HTTP response status codes and more. Based on PycURL and Python Requests."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
	"Programming Language :: Python :: 3",
	"License :: OSI Approved :: MIT License",
	"Operating System :: POSIX :: Linux"
]
dependencies = [
	"alive-progress>=3.1.5",
	"bot_safe_agents>=1.0",
	"colorama>=0.4.6",
	"pycurl>=7.45.2",
	"pyjwt>=2.7.0",
	"regex>=2023.8.8",
	"requests>=2.31.0",
	"tabulate>=0.9.0",
	"termcolor>=2.4.0"
]

[project.urls]
"Homepage" = "https://github.com/ivan-sincek/forbidden"

[project.scripts]
forbidden = "forbidden.main:main"
stresser = "stresser.main:main"

[tool.setuptools.packages.find]
where = ["src"]
